sample(auth): let each demo own its email link tag and move the routing helpers out of MainActivity
