+++ app/src/main/java/com/firebaseui/android/demo/MainActivity.kt
@@ -1,7 +1,6 @@
 package com.firebaseui.android.demo
 
 import android.content.Intent
-import android.net.Uri
 import android.os.Bundle
 import android.util.Log
 import androidx.activity.ComponentActivity
@@ -24,7 +23,6 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import androidx.core.net.toUri
 import com.firebase.ui.auth.FirebaseAuthUI
 import com.firebase.ui.auth.util.EmailLinkConstants
 import com.firebaseui.android.demo.auth.AuthChooserActivity
@@ -33,6 +31,7 @@ import com.firebaseui.android.demo.auth.fullcustomization.FullCustomizationDemoA
 import com.firebaseui.android.demo.database.DatabaseDemoActivity
 import com.firebaseui.android.demo.firestore.FirestoreDemoActivity
 import com.firebaseui.android.demo.storage.StorageDemoActivity
+import com.firebaseui.android.demo.utils.emailLinkOrigin
 import com.google.firebase.FirebaseApp
 import com.google.firebase.database.FirebaseDatabase
 import com.google.firebase.firestore.FirebaseFirestore
@@ -55,24 +54,6 @@ class MainActivity : ComponentActivity() {
         // useEmulator() throws once the Firestore/Database client has been used elsewhere in
         // the process, so this must only run once per process, not on every onCreate().
         private var emulatorsConfigured = false
-
-        // Path segments, not query params: ContinueUrlBuilder appends "?" and would corrupt a query.
-        const val ORIGIN_FULL_CUSTOMIZATION = "fullcustomization"
-        const val ORIGIN_HIGH_LEVEL = "highlevel"
-
-        /** The continue URL, which sits either directly on [uri] or nested inside its `link`. */
-        private fun continueUrlOf(uri: Uri): Uri? {
-            uri.getQueryParameter("continueUrl")?.let { return it.toUri() }
-            uri.getQueryParameter("link")?.let { return continueUrlOf(it.toUri()) }
-            return null
-        }
-
-        /** Which demo sent [link], or null when it says nothing about where it came from. */
-        internal fun emailLinkOrigin(link: String?): String? =
-            link?.takeIf { it.isNotEmpty() }
-                ?.let { runCatching { continueUrlOf(it.toUri()) }.getOrNull() }
-                ?.pathSegments
-                ?.lastOrNull()
     }
 
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -108,8 +89,9 @@ class MainActivity : ComponentActivity() {
 
         fun launchDemoForEmailLink() {
             val target = when (emailLinkOrigin(pendingEmailLink)) {
-                ORIGIN_FULL_CUSTOMIZATION -> FullCustomizationDemoActivity::class.java
-                // Untagged links predate this routing, so they belong to the demo that always handled them.
+                FullCustomizationDemoActivity.EMAIL_LINK_ORIGIN ->
+                    FullCustomizationDemoActivity::class.java
+                // Untagged links predate this routing, so they belong to the demo that had them.
                 else -> HighLevelApiDemoActivity::class.java
             }
             val demoIntent = Intent(this, target).apply {
