+++ app/src/main/java/com/firebaseui/android/demo/auth/HighLevelApiDemoActivity.kt
@@ -66,11 +66,15 @@ import com.firebase.ui.auth.ui.screens.reauth.ReauthContentState
 import com.firebase.ui.auth.util.EmailLinkConstants
 import com.firebase.ui.auth.util.displayIdentifier
 import com.firebase.ui.auth.util.getDisplayEmail
-import com.firebaseui.android.demo.MainActivity
 import com.firebaseui.android.demo.R
 import com.google.firebase.auth.actionCodeSettings
 
 class HighLevelApiDemoActivity : ComponentActivity() {
+    companion object {
+        /** Marks this demo's email links so MainActivity can route the return trip back here. */
+        const val EMAIL_LINK_ORIGIN = "highlevel"
+    }
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         enableEdgeToEdge()
@@ -132,9 +136,9 @@ class HighLevelApiDemoActivity : ComponentActivity() {
                             isEmailLinkForceSameDeviceEnabled = false,
                             isEmailLinkSignInEnabled = true,
                             emailLinkActionCodeSettings = actionCodeSettings {
-                                // Trailing segment is what MainActivity.emailLinkOrigin routes on.
+                                // Trailing segment is what MainActivity routes the return trip on.
                                 url = "https://flutterfire-e2e-tests.firebaseapp.com/demo/" +
-                                        MainActivity.ORIGIN_HIGH_LEVEL
+                                        EMAIL_LINK_ORIGIN
                                 handleCodeInApp = true
                                 setAndroidPackageName(
                                     "com.firebaseui.android.demo",
